requirements.txt
ari-py
websockets
aiohttp
pydantic
pydantic_settings
structlog
redis
PyYAML
vosk
soundfile
llama-cpp-python
piper-tts